Add files via upload
This chat shows Carolina Panther's Quarterback, Cameron Newton's career touchdowns and interceptions. This graph was made using Quartz Chartbuilder. Newton has been playing in the NFL since 2011 after graduating from Auburn University and entering the combine afterwards. The quarterback made it to the Superbowl in 2015 against the Denver Broncos but couldn't pull off the win. However, in that season he had the most TD's at 35. The team went 15-1 in the regular season in 2016.